HIVE/hive-backend-main/requirements.txt
Flask==2.2.3
google_api_python_client==2.84.0
google_auth_oauthlib==1.0.0
protobuf==4.22.3
pandas==1.5.3
txtai==5.4.0
slack_sdk==3.21.2
